(gabac) rename encode*, decode* with more specific names
