+++ src/genie/entropy/gabac/decode-desc-subseq.h
@@ -4,19 +4,19 @@
  * https://github.com/mitogen/genie for more details.
  */
 
-#ifndef GABAC_ENCODING_H_
-#define GABAC_ENCODING_H_
+#ifndef GABAC_DECODE_DESC_SUBSEQ_H_
+#define GABAC_DECODE_DESC_SUBSEQ_H_
 
 namespace genie {
 namespace entropy {
 namespace gabac {
 
-struct IOConfiguration;
 struct EncodingConfiguration;
+struct IOConfiguration;
 
-unsigned long encodeDescSubsequence(const IOConfiguration& conf, const EncodingConfiguration& enConf);
+unsigned long decodeDescSubsequence(const IOConfiguration& ioConf, const EncodingConfiguration& enConf);
 
 }  // namespace gabac
 }  // namespace entropy
 }  // namespace genie
-#endif  // GABAC_ENCODING_H_
+#endif  // GABAC_DECODE_DESC_SUBSEQ_H_
