+++ src/genie/entropy/gabac/decode-transformed-subseq.cc
@@ -4,14 +4,15 @@
  * https://github.com/mitogen/genie for more details.
  */
 
-#include "encode-cabac.h"
+#include "decode-transformed-subseq.h"
 
 #include <cassert>
+#include <limits>
 
-#include <genie/util/block-stepper.h>
 #include <genie/util/data-block.h>
+#include <genie/entropy/paramcabac/subsequence.h>
 #include "exceptions.h"
-#include "writer.h"
+#include "reader.h"
 
 #include "context-selector.h"
 #include "luts-subsymbol-transform.h"
@@ -20,19 +21,22 @@ namespace genie {
 namespace entropy {
 namespace gabac {
 
-typedef void (Writer::*binFunc)(uint64_t, const std::vector<unsigned int>);
+typedef uint64_t (Reader::*binFunc)(const std::vector<unsigned int>);
 
 static inline
-void encodeSignflag(Writer &writer,
+void decodeSignFlag(Reader &reader,
                     const paramcabac::BinarizationParameters::BinarizationId binID,
-                    const int64_t signedSymbolValue) {
-    if(signedSymbolValue != 0) {
+                    uint64_t &symbolValue) {
+    if(symbolValue != 0) {
         switch(binID) {
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_EXPONENTIAL_GOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_TRUNCATED_EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_SPLIT_UNITWISE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_DOUBLE_TRUNCATED_UNARY:
-                writer.writeSignFlag(signedSymbolValue);
+                if(reader.readSignFlag()) {
+                    int64_t symbolValueSigned = -(symbolValue);
+                    symbolValue = static_cast<uint64_t>(symbolValueSigned);
+                }
                 break;
             default:
                 break;
@@ -51,31 +55,31 @@ binFunc getBinarizor(const uint8_t outputSymbolSize,
     if(bypassFlag) {
         switch (binID) {
             case paramcabac::BinarizationParameters::BinarizationId::BINARY_CODING:
-                func = &Writer::writeAsBIbypass;
+                func = &Reader::readAsBIbypass;
                 binParams[0] = stateVars.getCLengthBI();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY:
-                func = &Writer::writeAsTUbypass;
+                func = &Reader::readAsTUbypass;
                 binParams[0] = binarzationParams.getCMax();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_EXPONENTIAL_GOMB:
-                func = &Writer::writeAsEGbypass;
+                func = &Reader::readAsEGbypass;
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_TRUNCATED_EXPONENTIAL_GOLOMB:
-                func = &Writer::writeAsTEGbypass;
+                func = &Reader::readAsTEGbypass;
                 binParams[0] = binarzationParams.getCMaxTeg();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::SPLIT_UNITWISE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_SPLIT_UNITWISE_TRUNCATED_UNARY:
-                func = &Writer::writeAsSUTUbypass;
+                func = &Reader::readAsSUTUbypass;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::DOUBLE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_DOUBLE_TRUNCATED_UNARY:
-                func = &Writer::writeAsDTUbypass;
+                func = &Reader::readAsDTUbypass;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 binParams[2] = binarzationParams.getCMaxDtu();
@@ -86,31 +90,31 @@ binFunc getBinarizor(const uint8_t outputSymbolSize,
     } else {
         switch (binID) {
             case paramcabac::BinarizationParameters::BinarizationId::BINARY_CODING:
-                func = &Writer::writeAsBIcabac;
+                func = &Reader::readAsBIcabac;
                 binParams[0] = stateVars.getCLengthBI();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY:
-                func = &Writer::writeAsTUcabac;
+                func = &Reader::readAsTUcabac;
                 binParams[0] = binarzationParams.getCMax();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_EXPONENTIAL_GOMB:
-                func = &Writer::writeAsEGcabac;
+                func = &Reader::readAsEGcabac;
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_TRUNCATED_EXPONENTIAL_GOLOMB:
-                func = &Writer::writeAsTEGcabac;
+                func = &Reader::readAsTEGcabac;
                 binParams[0] = binarzationParams.getCMaxTeg();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::SPLIT_UNITWISE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_SPLIT_UNITWISE_TRUNCATED_UNARY:
-                func = &Writer::writeAsSUTUcabac;
+                func = &Reader::readAsSUTUcabac;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::DOUBLE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_DOUBLE_TRUNCATED_UNARY:
-                func = &Writer::writeAsDTUcabac;
+                func = &Reader::readAsDTUcabac;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 binParams[2] = binarzationParams.getCMaxDtu();
@@ -123,13 +127,14 @@ binFunc getBinarizor(const uint8_t outputSymbolSize,
     return func;
 }
 
-size_t encodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                                   util::DataBlock* symbols,
-                                   size_t maxSize) {
-    assert(symbols != nullptr);
+size_t decodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                                   const unsigned int numEncodedSymbols,
+                                   util::DataBlock* bitstream) {
+    if (bitstream == nullptr) {
+        GABAC_DIE("Bitstream is null");
+    }
 
-    size_t numSymbols = symbols->size();
-    if (numSymbols <= 0) return 0;
+    if (numEncodedSymbols <= 0) return 0;
 
     const paramcabac::SupportValues &supportVals = trnsfSubseqConf.getSupportValues();
     const paramcabac::Binarization &binarzation = trnsfSubseqConf.getBinarization();
@@ -139,19 +144,18 @@ size_t encodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSub
 
     const uint8_t outputSymbolSize = supportVals.getOutputSymbolSize();
     const uint8_t codingSubsymSize = supportVals.getCodingSubsymSize();
-    const uint64_t subsymMask = paramcabac::StateVars::get2PowN(codingSubsymSize)-1;
     const bool bypassFlag = binarzation.getBypassFlag();
+    size_t payloadSizeUsed = 0;
 
-    util::DataBlock block(0, 1);
-    OBufferStream bitstream(&block);
-    Writer writer(&bitstream,
+    Reader reader(bitstream,
                   bypassFlag,
                   stateVars.getNumCtxTotal());
-    writer.start();
+    reader.start();
 
-    std::vector<unsigned int> binParams(4, 0);  // first three elements are for binarization params, last one is for ctxIdx
+    std::vector<unsigned int> binParams(4, 0); // first three elements are for binarization params, last one is for ctxIdx
 
-    util::BlockStepper r = symbols->getReader();
+    util::DataBlock decodedSymbols(numEncodedSymbols, 4);
+    util::BlockStepper r = decodedSymbols.getReader();
     std::vector<Subsymbol> subsymbols(stateVars.getNumSubsymbols());
 
     ContextSelector ctxSelector(stateVars);
@@ -165,53 +169,46 @@ size_t encodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSub
                                 binParams);
 
     while (r.isValid()) {
-        if (maxSize <= bitstream.size()) {
-            break;
-        }
+        // Decode subsymbols and merge them to construct symbols
+        uint64_t symbolValue = 0;
 
-        // Split symbol into subsymbols and then encode subsymbols
-        const uint64_t origSymbol = r.get();
-        const int64_t signedSymbolValue = paramcabac::StateVars::getSignedValue(origSymbol, symbols->getWordSize());
-        const uint64_t symbolValue = abs(signedSymbolValue);
-        uint64_t subsymValToCode = 0;
-
-        uint32_t oss = outputSymbolSize;
         for (uint8_t s=0; s<stateVars.getNumSubsymbols(); s++) {
 
-            subsymValToCode = subsymbols[s].subsymValue = (symbolValue>>(oss-=codingSubsymSize)) & subsymMask;
             subsymbols[s].subsymIdx = s;
+            binParams[3] = ctxSelector.getContextIdxOrder0(s);
+
+            subsymbols[s].subsymValue = (reader.*func)(binParams);
 
             if(diffEnabled) {
-                assert(subsymbols[s].prvValues[0] <= subsymValToCode);
-                subsymValToCode -= subsymbols[s].prvValues[0];
+                subsymbols[s].subsymValue += subsymbols[s].prvValues[0];
                 subsymbols[s].prvValues[0] = subsymbols[s].subsymValue;
             }
 
-            binParams[3] = ctxSelector.getContextIdxOrder0(s);
-
-            (writer.*func)(subsymValToCode, binParams);
+            symbolValue = (symbolValue<<codingSubsymSize) | subsymbols[s].subsymValue;
         }
 
-        encodeSignflag(writer, binID, signedSymbolValue);
+        decodeSignFlag(reader, binID, symbolValue);
 
+        r.set(symbolValue);
         r.inc();
     }
 
-    writer.close();
+    payloadSizeUsed = reader.close();
 
-    bitstream.flush(symbols);
+    decodedSymbols.swap(bitstream);
 
-    return symbols->size(); // size of bitstream
+    return payloadSizeUsed;
 }
 
-size_t encodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                                   util::DataBlock* symbols,
-                                   util::DataBlock* const depSymbols,
-                                   size_t maxSize) {
-    assert(symbols != nullptr);
+size_t decodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                                   const unsigned int numEncodedSymbols,
+                                   util::DataBlock* bitstream,
+                                   util::DataBlock* const depSymbols) {
+    if (bitstream == nullptr) {
+        GABAC_DIE("Bitstream is null");
+    }
 
-    size_t numSymbols = symbols->size();
-    if (numSymbols <= 0) return 0;
+    if (numEncodedSymbols <= 0) return 0;
 
     const paramcabac::SupportValues &supportVals = trnsfSubseqConf.getSupportValues();
     const paramcabac::Binarization &binarzation = trnsfSubseqConf.getBinarization();
@@ -231,22 +228,26 @@ size_t encodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSub
                                                  supportVals.getShareSubsymLutFlag(),
                                                  trnsfSubseqConf.getTransformIDSubsym());
     uint8_t const numPrvs = stateVars.getNumPrvs(supportVals.getShareSubsymPrvFlag());
+    size_t payloadSizeUsed = 0;
 
-    util::DataBlock block(0, 1);
-    OBufferStream bitstream(&block);
-    Writer writer(&bitstream,
+    Reader reader(bitstream,
                   bypassFlag,
                   stateVars.getNumCtxTotal());
-    writer.start();
+    reader.start();
 
     std::vector<unsigned int> binParams(4, 0); // first three elements are for binarization params, last one is for ctxIdx
 
-    util::BlockStepper r = symbols->getReader();
+    util::DataBlock decodedSymbols(numEncodedSymbols, 4);
+    util::BlockStepper r = decodedSymbols.getReader();
     std::vector<Subsymbol> subsymbols(stateVars.getNumSubsymbols());
 
-    LUTsSubSymbolTransform lutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, true);
+    LUTsSubSymbolTransform invLutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, false);
+    bool customCmaxTU = false;
     if(numLuts > 0) {
-        lutsSubsymTrnsfm.encodeLUTs(writer, alphaProps, symbols, depSymbols);
+        invLutsSubsymTrnsfm.decodeLUTs(reader);
+        if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
+            customCmaxTU = true;
+        }
     }
 
     util::BlockStepper rDep;
@@ -264,15 +265,8 @@ size_t encodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSub
                                 binParams);
 
     while (r.isValid()) {
-        if (maxSize <= bitstream.size()) {
-            break;
-        }
-
-        // Split symbol into subsymbols and then encode subsymbols
-        const uint64_t origSymbol = r.get();
-        const int64_t signedSymbolValue = paramcabac::StateVars::getSignedValue(origSymbol, symbols->getWordSize());
-        const uint64_t symbolValue = abs(signedSymbolValue);
-        uint64_t subsymValToCode = 0;
+        // Decode subsymbols and merge them to construct symbols
+        uint64_t symbolValue = 0;
 
         uint64_t depSymbolValue = 0, depSubsymValue = 0;
         if(rDep.isValid()) {
@@ -286,63 +280,65 @@ size_t encodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSub
             const uint8_t prvIdx = (numPrvs > 1) ? s : 0; // either private or shared PRV
 
             if(depSymbols) {
-                depSubsymValue = (depSymbolValue>>(oss-codingSubsymSize)) & subsymMask;
+                depSubsymValue = (depSymbolValue>>(oss-=codingSubsymSize)) & subsymMask;
                 subsymbols[prvIdx].prvValues[0] = depSubsymValue;
             }
 
-            subsymValToCode = subsymbols[s].subsymValue = (symbolValue>>(oss-=codingSubsymSize)) & subsymMask;
             subsymbols[s].subsymIdx = s;
-
             binParams[3] = ctxSelector.getContextIdxOrderGT0(s,
                                                              prvIdx,
                                                              subsymbols,
                                                              codingOrder);
 
+            if(customCmaxTU) {
+                subsymbols[s].lutNumMaxElems = invLutsSubsymTrnsfm.getNumMaxElemsOrder1(subsymbols,
+                                                                                        lutIdx,
+                                                                                        prvIdx);
+                binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
+            }
+            subsymbols[s].subsymValue = (reader.*func)(binParams);
+
             if(numLuts > 0) {
-                subsymbols[s].lutEntryIdx = 0;
-                lutsSubsymTrnsfm.transformOrder1(subsymbols, s, lutIdx, prvIdx);
-                subsymValToCode = subsymbols[s].lutEntryIdx;
-                if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
-                    binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
-                }
+                subsymbols[s].lutEntryIdx = subsymbols[s].subsymValue;
+                invLutsSubsymTrnsfm.invTransformOrder1(subsymbols, s, lutIdx, prvIdx);
             }
 
-            (writer.*func)(subsymValToCode, binParams);
-
             subsymbols[prvIdx].prvValues[0] = subsymbols[s].subsymValue;
+
+            symbolValue = (symbolValue<<codingSubsymSize) | subsymbols[s].subsymValue;
         }
 
-        encodeSignflag(writer, binID, signedSymbolValue);
+        decodeSignFlag(reader, binID, symbolValue);
 
+        r.set(symbolValue);
         r.inc();
     }
 
-    writer.close();
+    payloadSizeUsed = reader.close();
 
-    bitstream.flush(symbols);
+    decodedSymbols.swap(bitstream);
 
-    return symbols->size(); // size of bitstream
+    return payloadSizeUsed;
 }
 
-size_t encodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                                   util::DataBlock* symbols,
-                                   size_t maxSize) {
-    assert(symbols != nullptr);
+size_t decodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                                   const unsigned int numEncodedSymbols,
+                                   util::DataBlock* bitstream) {
+    if (bitstream == nullptr) {
+        GABAC_DIE("Bitstream is null");
+    }
 
-    size_t numSymbols = symbols->size();
-    if (numSymbols <= 0) return 0;
+    if (numEncodedSymbols <= 0) return 0;
 
     const paramcabac::SupportValues &supportVals = trnsfSubseqConf.getSupportValues();
     const paramcabac::Binarization &binarzation = trnsfSubseqConf.getBinarization();
     const paramcabac::BinarizationParameters &binarzationParams = binarzation.getCabacBinarizationParameters();
     const paramcabac::StateVars &stateVars = trnsfSubseqConf.getStateVars();
     const paramcabac::BinarizationParameters::BinarizationId binID = binarzation.getBinarizationID();
-    const core::Alphabet alphaProps = getAlphabetProperties(trnsfSubseqConf.getAlphabetID());
 
     const uint8_t outputSymbolSize = supportVals.getOutputSymbolSize();
     const uint8_t codingSubsymSize = supportVals.getCodingSubsymSize();
     const uint8_t codingOrder = supportVals.getCodingOrder();
-    const uint64_t subsymMask = paramcabac::StateVars::get2PowN(codingSubsymSize)-1;
     const bool bypassFlag = binarzation.getBypassFlag();
     assert(bypassFlag == false);
 
@@ -350,22 +346,26 @@ size_t encodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSub
                                                  supportVals.getShareSubsymLutFlag(),
                                                  trnsfSubseqConf.getTransformIDSubsym());
     uint8_t const numPrvs = stateVars.getNumPrvs(supportVals.getShareSubsymPrvFlag());
+    size_t payloadSizeUsed = 0;
 
-    util::DataBlock block(0, 1);
-    OBufferStream bitstream(&block);
-    Writer writer(&bitstream,
+    Reader reader(bitstream,
                   bypassFlag,
                   stateVars.getNumCtxTotal());
-    writer.start();
+    reader.start();
 
     std::vector<unsigned int> binParams(4, 0); // first three elements are for binarization params, last one is for ctxIdx
 
-    util::BlockStepper r = symbols->getReader();
+    util::DataBlock decodedSymbols(numEncodedSymbols, 4);
+    util::BlockStepper r = decodedSymbols.getReader();
     std::vector<Subsymbol> subsymbols(stateVars.getNumSubsymbols());
 
-    LUTsSubSymbolTransform lutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, true);
+    LUTsSubSymbolTransform invLutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, false);
+    bool customCmaxTU = false;
     if(numLuts > 0) {
-        lutsSubsymTrnsfm.encodeLUTs(writer, alphaProps, symbols, nullptr);
+        invLutsSubsymTrnsfm.decodeLUTs(reader);
+        if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
+            customCmaxTU = true;
+        }
     }
 
     ContextSelector ctxSelector(stateVars);
@@ -378,70 +378,64 @@ size_t encodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSub
                                 binParams);
 
     while (r.isValid()) {
-        if (maxSize <= bitstream.size()) {
-            break;
-        }
+        // Decode subsymbols and merge them to construct symbols
+        uint64_t symbolValue = 0;
 
-        // Split symbol into subsymbols and then encode subsymbols
-        const uint64_t origSymbol = r.get();
-        const int64_t signedSymbolValue = paramcabac::StateVars::getSignedValue(origSymbol, symbols->getWordSize());
-        const uint64_t symbolValue = abs(signedSymbolValue);
-        uint64_t subsymValToCode = 0;
-
-        uint32_t oss = outputSymbolSize;
         for (uint8_t s=0; s<stateVars.getNumSubsymbols(); s++) {
             const uint8_t lutIdx = (numLuts > 1) ? s : 0; // either private or shared LUT
             const uint8_t prvIdx = (numPrvs > 1) ? s : 0; // either private or shared PRV
 
-            subsymValToCode = subsymbols[s].subsymValue = (symbolValue>>(oss-=codingSubsymSize)) & subsymMask;
             subsymbols[s].subsymIdx = s;
-
             binParams[3] = ctxSelector.getContextIdxOrderGT0(s,
                                                              prvIdx,
                                                              subsymbols,
                                                              codingOrder);
 
-            if(numLuts > 0) {
-                subsymbols[s].lutEntryIdx = 0;
-                lutsSubsymTrnsfm.transformOrder2(subsymbols, s, lutIdx, prvIdx);
-                subsymValToCode = subsymbols[s].lutEntryIdx;
-                if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
-                    binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
-                }
+            if(customCmaxTU) {
+                subsymbols[s].lutNumMaxElems = invLutsSubsymTrnsfm.getNumMaxElemsOrder2(subsymbols,
+                                                                                        lutIdx,
+                                                                                        prvIdx);
+                binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
             }
+            subsymbols[s].subsymValue = (reader.*func)(binParams);
 
-            (writer.*func)(subsymValToCode, binParams);
+            if(numLuts > 0) {
+                subsymbols[s].lutEntryIdx = subsymbols[s].subsymValue;
+                invLutsSubsymTrnsfm.invTransformOrder2(subsymbols, s, lutIdx, prvIdx);
+            }
 
             subsymbols[prvIdx].prvValues[1] = subsymbols[prvIdx].prvValues[0];
             subsymbols[prvIdx].prvValues[0] = subsymbols[s].subsymValue;
+
+            symbolValue = (symbolValue<<codingSubsymSize) | subsymbols[s].subsymValue;
         }
 
-        encodeSignflag(writer, binID, signedSymbolValue);
+        decodeSignFlag(reader, binID, symbolValue);
 
+        r.set(symbolValue);
         r.inc();
     }
 
-    writer.close();
+    payloadSizeUsed = reader.close();
 
-    bitstream.flush(symbols);
+    decodedSymbols.swap(bitstream);
 
-    return symbols->size(); // size of bitstream
+    return payloadSizeUsed;
 }
 
-size_t encodeTransformSubseq(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                             util::DataBlock* symbols,
-                             util::DataBlock* const depSymbols,
-                             size_t maxSize) {
-
+size_t decodeTransformSubseq(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                             const unsigned int numEncodedSymbols,
+                             util::DataBlock* bitstream,
+                             util::DataBlock* const depSymbols) {
     switch(trnsfSubseqConf.getSupportValues().getCodingOrder()) {
         case 0:
-            return encodeTransformSubseqOrder0(trnsfSubseqConf, symbols, maxSize);
+            return decodeTransformSubseqOrder0(trnsfSubseqConf, numEncodedSymbols, bitstream);
         break;
         case 1:
-            return encodeTransformSubseqOrder1(trnsfSubseqConf, symbols, depSymbols, maxSize);
+            return decodeTransformSubseqOrder1(trnsfSubseqConf, numEncodedSymbols, bitstream, depSymbols);
         break;
         case 2:
-            return encodeTransformSubseqOrder2(trnsfSubseqConf, symbols, maxSize);
+            return decodeTransformSubseqOrder2(trnsfSubseqConf, numEncodedSymbols, bitstream);
         break;
         default:
             GABAC_DIE("Unknown coding order");
