+++ src/genie/entropy/gabac/encode-transformed-subseq.cc
@@ -4,15 +4,14 @@
  * https://github.com/mitogen/genie for more details.
  */
 
-#include "decode-cabac.h"
+#include "encode-transformed-subseq.h"
 
 #include <cassert>
-#include <limits>
 
+#include <genie/util/block-stepper.h>
 #include <genie/util/data-block.h>
-#include <genie/entropy/paramcabac/subsequence.h>
 #include "exceptions.h"
-#include "reader.h"
+#include "writer.h"
 
 #include "context-selector.h"
 #include "luts-subsymbol-transform.h"
@@ -21,22 +20,19 @@ namespace genie {
 namespace entropy {
 namespace gabac {
 
-typedef uint64_t (Reader::*binFunc)(const std::vector<unsigned int>);
+typedef void (Writer::*binFunc)(uint64_t, const std::vector<unsigned int>);
 
 static inline
-void decodeSignFlag(Reader &reader,
+void encodeSignflag(Writer &writer,
                     const paramcabac::BinarizationParameters::BinarizationId binID,
-                    uint64_t &symbolValue) {
-    if(symbolValue != 0) {
+                    const int64_t signedSymbolValue) {
+    if(signedSymbolValue != 0) {
         switch(binID) {
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_EXPONENTIAL_GOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_TRUNCATED_EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_SPLIT_UNITWISE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_DOUBLE_TRUNCATED_UNARY:
-                if(reader.readSignFlag()) {
-                    int64_t symbolValueSigned = -(symbolValue);
-                    symbolValue = static_cast<uint64_t>(symbolValueSigned);
-                }
+                writer.writeSignFlag(signedSymbolValue);
                 break;
             default:
                 break;
@@ -55,31 +51,31 @@ binFunc getBinarizor(const uint8_t outputSymbolSize,
     if(bypassFlag) {
         switch (binID) {
             case paramcabac::BinarizationParameters::BinarizationId::BINARY_CODING:
-                func = &Reader::readAsBIbypass;
+                func = &Writer::writeAsBIbypass;
                 binParams[0] = stateVars.getCLengthBI();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY:
-                func = &Reader::readAsTUbypass;
+                func = &Writer::writeAsTUbypass;
                 binParams[0] = binarzationParams.getCMax();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_EXPONENTIAL_GOMB:
-                func = &Reader::readAsEGbypass;
+                func = &Writer::writeAsEGbypass;
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_TRUNCATED_EXPONENTIAL_GOLOMB:
-                func = &Reader::readAsTEGbypass;
+                func = &Writer::writeAsTEGbypass;
                 binParams[0] = binarzationParams.getCMaxTeg();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::SPLIT_UNITWISE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_SPLIT_UNITWISE_TRUNCATED_UNARY:
-                func = &Reader::readAsSUTUbypass;
+                func = &Writer::writeAsSUTUbypass;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::DOUBLE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_DOUBLE_TRUNCATED_UNARY:
-                func = &Reader::readAsDTUbypass;
+                func = &Writer::writeAsDTUbypass;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 binParams[2] = binarzationParams.getCMaxDtu();
@@ -90,31 +86,31 @@ binFunc getBinarizor(const uint8_t outputSymbolSize,
     } else {
         switch (binID) {
             case paramcabac::BinarizationParameters::BinarizationId::BINARY_CODING:
-                func = &Reader::readAsBIcabac;
+                func = &Writer::writeAsBIcabac;
                 binParams[0] = stateVars.getCLengthBI();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY:
-                func = &Reader::readAsTUcabac;
+                func = &Writer::writeAsTUcabac;
                 binParams[0] = binarzationParams.getCMax();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_EXPONENTIAL_GOMB:
-                func = &Reader::readAsEGcabac;
+                func = &Writer::writeAsEGcabac;
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_EXPONENTIAL_GOLOMB:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_TRUNCATED_EXPONENTIAL_GOLOMB:
-                func = &Reader::readAsTEGcabac;
+                func = &Writer::writeAsTEGcabac;
                 binParams[0] = binarzationParams.getCMaxTeg();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::SPLIT_UNITWISE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_SPLIT_UNITWISE_TRUNCATED_UNARY:
-                func = &Reader::readAsSUTUcabac;
+                func = &Writer::writeAsSUTUcabac;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 break;
             case paramcabac::BinarizationParameters::BinarizationId::DOUBLE_TRUNCATED_UNARY:
             case paramcabac::BinarizationParameters::BinarizationId::SIGNED_DOUBLE_TRUNCATED_UNARY:
-                func = &Reader::readAsDTUcabac;
+                func = &Writer::writeAsDTUcabac;
                 binParams[0] = outputSymbolSize;
                 binParams[1] = binarzationParams.getSplitUnitSize();
                 binParams[2] = binarzationParams.getCMaxDtu();
@@ -127,14 +123,13 @@ binFunc getBinarizor(const uint8_t outputSymbolSize,
     return func;
 }
 
-size_t decodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                                   const unsigned int numEncodedSymbols,
-                                   util::DataBlock* bitstream) {
-    if (bitstream == nullptr) {
-        GABAC_DIE("Bitstream is null");
-    }
+size_t encodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                                   util::DataBlock* symbols,
+                                   size_t maxSize) {
+    assert(symbols != nullptr);
 
-    if (numEncodedSymbols <= 0) return 0;
+    size_t numSymbols = symbols->size();
+    if (numSymbols <= 0) return 0;
 
     const paramcabac::SupportValues &supportVals = trnsfSubseqConf.getSupportValues();
     const paramcabac::Binarization &binarzation = trnsfSubseqConf.getBinarization();
@@ -144,18 +139,19 @@ size_t decodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSub
 
     const uint8_t outputSymbolSize = supportVals.getOutputSymbolSize();
     const uint8_t codingSubsymSize = supportVals.getCodingSubsymSize();
+    const uint64_t subsymMask = paramcabac::StateVars::get2PowN(codingSubsymSize)-1;
     const bool bypassFlag = binarzation.getBypassFlag();
-    size_t payloadSizeUsed = 0;
 
-    Reader reader(bitstream,
+    util::DataBlock block(0, 1);
+    OBufferStream bitstream(&block);
+    Writer writer(&bitstream,
                   bypassFlag,
                   stateVars.getNumCtxTotal());
-    reader.start();
+    writer.start();
 
-    std::vector<unsigned int> binParams(4, 0); // first three elements are for binarization params, last one is for ctxIdx
+    std::vector<unsigned int> binParams(4, 0);  // first three elements are for binarization params, last one is for ctxIdx
 
-    util::DataBlock decodedSymbols(numEncodedSymbols, 4);
-    util::BlockStepper r = decodedSymbols.getReader();
+    util::BlockStepper r = symbols->getReader();
     std::vector<Subsymbol> subsymbols(stateVars.getNumSubsymbols());
 
     ContextSelector ctxSelector(stateVars);
@@ -169,46 +165,53 @@ size_t decodeTransformSubseqOrder0(const paramcabac::TransformedSubSeq &trnsfSub
                                 binParams);
 
     while (r.isValid()) {
-        // Decode subsymbols and merge them to construct symbols
-        uint64_t symbolValue = 0;
+        if (maxSize <= bitstream.size()) {
+            break;
+        }
 
+        // Split symbol into subsymbols and then encode subsymbols
+        const uint64_t origSymbol = r.get();
+        const int64_t signedSymbolValue = paramcabac::StateVars::getSignedValue(origSymbol, symbols->getWordSize());
+        const uint64_t symbolValue = abs(signedSymbolValue);
+        uint64_t subsymValToCode = 0;
+
+        uint32_t oss = outputSymbolSize;
         for (uint8_t s=0; s<stateVars.getNumSubsymbols(); s++) {
 
+            subsymValToCode = subsymbols[s].subsymValue = (symbolValue>>(oss-=codingSubsymSize)) & subsymMask;
             subsymbols[s].subsymIdx = s;
-            binParams[3] = ctxSelector.getContextIdxOrder0(s);
-
-            subsymbols[s].subsymValue = (reader.*func)(binParams);
 
             if(diffEnabled) {
-                subsymbols[s].subsymValue += subsymbols[s].prvValues[0];
+                assert(subsymbols[s].prvValues[0] <= subsymValToCode);
+                subsymValToCode -= subsymbols[s].prvValues[0];
                 subsymbols[s].prvValues[0] = subsymbols[s].subsymValue;
             }
 
-            symbolValue = (symbolValue<<codingSubsymSize) | subsymbols[s].subsymValue;
+            binParams[3] = ctxSelector.getContextIdxOrder0(s);
+
+            (writer.*func)(subsymValToCode, binParams);
         }
 
-        decodeSignFlag(reader, binID, symbolValue);
+        encodeSignflag(writer, binID, signedSymbolValue);
 
-        r.set(symbolValue);
         r.inc();
     }
 
-    payloadSizeUsed = reader.close();
+    writer.close();
 
-    decodedSymbols.swap(bitstream);
+    bitstream.flush(symbols);
 
-    return payloadSizeUsed;
+    return symbols->size(); // size of bitstream
 }
 
-size_t decodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                                   const unsigned int numEncodedSymbols,
-                                   util::DataBlock* bitstream,
-                                   util::DataBlock* const depSymbols) {
-    if (bitstream == nullptr) {
-        GABAC_DIE("Bitstream is null");
-    }
+size_t encodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                                   util::DataBlock* symbols,
+                                   util::DataBlock* const depSymbols,
+                                   size_t maxSize) {
+    assert(symbols != nullptr);
 
-    if (numEncodedSymbols <= 0) return 0;
+    size_t numSymbols = symbols->size();
+    if (numSymbols <= 0) return 0;
 
     const paramcabac::SupportValues &supportVals = trnsfSubseqConf.getSupportValues();
     const paramcabac::Binarization &binarzation = trnsfSubseqConf.getBinarization();
@@ -228,26 +231,22 @@ size_t decodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSub
                                                  supportVals.getShareSubsymLutFlag(),
                                                  trnsfSubseqConf.getTransformIDSubsym());
     uint8_t const numPrvs = stateVars.getNumPrvs(supportVals.getShareSubsymPrvFlag());
-    size_t payloadSizeUsed = 0;
 
-    Reader reader(bitstream,
+    util::DataBlock block(0, 1);
+    OBufferStream bitstream(&block);
+    Writer writer(&bitstream,
                   bypassFlag,
                   stateVars.getNumCtxTotal());
-    reader.start();
+    writer.start();
 
     std::vector<unsigned int> binParams(4, 0); // first three elements are for binarization params, last one is for ctxIdx
 
-    util::DataBlock decodedSymbols(numEncodedSymbols, 4);
-    util::BlockStepper r = decodedSymbols.getReader();
+    util::BlockStepper r = symbols->getReader();
     std::vector<Subsymbol> subsymbols(stateVars.getNumSubsymbols());
 
-    LUTsSubSymbolTransform invLutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, false);
-    bool customCmaxTU = false;
+    LUTsSubSymbolTransform lutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, true);
     if(numLuts > 0) {
-        invLutsSubsymTrnsfm.decodeLUTs(reader);
-        if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
-            customCmaxTU = true;
-        }
+        lutsSubsymTrnsfm.encodeLUTs(writer, alphaProps, symbols, depSymbols);
     }
 
     util::BlockStepper rDep;
@@ -265,8 +264,15 @@ size_t decodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSub
                                 binParams);
 
     while (r.isValid()) {
-        // Decode subsymbols and merge them to construct symbols
-        uint64_t symbolValue = 0;
+        if (maxSize <= bitstream.size()) {
+            break;
+        }
+
+        // Split symbol into subsymbols and then encode subsymbols
+        const uint64_t origSymbol = r.get();
+        const int64_t signedSymbolValue = paramcabac::StateVars::getSignedValue(origSymbol, symbols->getWordSize());
+        const uint64_t symbolValue = abs(signedSymbolValue);
+        uint64_t subsymValToCode = 0;
 
         uint64_t depSymbolValue = 0, depSubsymValue = 0;
         if(rDep.isValid()) {
@@ -280,65 +286,63 @@ size_t decodeTransformSubseqOrder1(const paramcabac::TransformedSubSeq &trnsfSub
             const uint8_t prvIdx = (numPrvs > 1) ? s : 0; // either private or shared PRV
 
             if(depSymbols) {
-                depSubsymValue = (depSymbolValue>>(oss-=codingSubsymSize)) & subsymMask;
+                depSubsymValue = (depSymbolValue>>(oss-codingSubsymSize)) & subsymMask;
                 subsymbols[prvIdx].prvValues[0] = depSubsymValue;
             }
 
+            subsymValToCode = subsymbols[s].subsymValue = (symbolValue>>(oss-=codingSubsymSize)) & subsymMask;
             subsymbols[s].subsymIdx = s;
+
             binParams[3] = ctxSelector.getContextIdxOrderGT0(s,
                                                              prvIdx,
                                                              subsymbols,
                                                              codingOrder);
 
-            if(customCmaxTU) {
-                subsymbols[s].lutNumMaxElems = invLutsSubsymTrnsfm.getNumMaxElemsOrder1(subsymbols,
-                                                                                        lutIdx,
-                                                                                        prvIdx);
-                binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
-            }
-            subsymbols[s].subsymValue = (reader.*func)(binParams);
-
             if(numLuts > 0) {
-                subsymbols[s].lutEntryIdx = subsymbols[s].subsymValue;
-                invLutsSubsymTrnsfm.invTransformOrder1(subsymbols, s, lutIdx, prvIdx);
+                subsymbols[s].lutEntryIdx = 0;
+                lutsSubsymTrnsfm.transformOrder1(subsymbols, s, lutIdx, prvIdx);
+                subsymValToCode = subsymbols[s].lutEntryIdx;
+                if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
+                    binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
+                }
             }
+
+            (writer.*func)(subsymValToCode, binParams);
 
             subsymbols[prvIdx].prvValues[0] = subsymbols[s].subsymValue;
-
-            symbolValue = (symbolValue<<codingSubsymSize) | subsymbols[s].subsymValue;
         }
 
-        decodeSignFlag(reader, binID, symbolValue);
+        encodeSignflag(writer, binID, signedSymbolValue);
 
-        r.set(symbolValue);
         r.inc();
     }
 
-    payloadSizeUsed = reader.close();
+    writer.close();
 
-    decodedSymbols.swap(bitstream);
+    bitstream.flush(symbols);
 
-    return payloadSizeUsed;
+    return symbols->size(); // size of bitstream
 }
 
-size_t decodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                                   const unsigned int numEncodedSymbols,
-                                   util::DataBlock* bitstream) {
-    if (bitstream == nullptr) {
-        GABAC_DIE("Bitstream is null");
-    }
+size_t encodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                                   util::DataBlock* symbols,
+                                   size_t maxSize) {
+    assert(symbols != nullptr);
 
-    if (numEncodedSymbols <= 0) return 0;
+    size_t numSymbols = symbols->size();
+    if (numSymbols <= 0) return 0;
 
     const paramcabac::SupportValues &supportVals = trnsfSubseqConf.getSupportValues();
     const paramcabac::Binarization &binarzation = trnsfSubseqConf.getBinarization();
     const paramcabac::BinarizationParameters &binarzationParams = binarzation.getCabacBinarizationParameters();
     const paramcabac::StateVars &stateVars = trnsfSubseqConf.getStateVars();
     const paramcabac::BinarizationParameters::BinarizationId binID = binarzation.getBinarizationID();
+    const core::Alphabet alphaProps = getAlphabetProperties(trnsfSubseqConf.getAlphabetID());
 
     const uint8_t outputSymbolSize = supportVals.getOutputSymbolSize();
     const uint8_t codingSubsymSize = supportVals.getCodingSubsymSize();
     const uint8_t codingOrder = supportVals.getCodingOrder();
+    const uint64_t subsymMask = paramcabac::StateVars::get2PowN(codingSubsymSize)-1;
     const bool bypassFlag = binarzation.getBypassFlag();
     assert(bypassFlag == false);
 
@@ -346,26 +350,22 @@ size_t decodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSub
                                                  supportVals.getShareSubsymLutFlag(),
                                                  trnsfSubseqConf.getTransformIDSubsym());
     uint8_t const numPrvs = stateVars.getNumPrvs(supportVals.getShareSubsymPrvFlag());
-    size_t payloadSizeUsed = 0;
 
-    Reader reader(bitstream,
+    util::DataBlock block(0, 1);
+    OBufferStream bitstream(&block);
+    Writer writer(&bitstream,
                   bypassFlag,
                   stateVars.getNumCtxTotal());
-    reader.start();
+    writer.start();
 
     std::vector<unsigned int> binParams(4, 0); // first three elements are for binarization params, last one is for ctxIdx
 
-    util::DataBlock decodedSymbols(numEncodedSymbols, 4);
-    util::BlockStepper r = decodedSymbols.getReader();
+    util::BlockStepper r = symbols->getReader();
     std::vector<Subsymbol> subsymbols(stateVars.getNumSubsymbols());
 
-    LUTsSubSymbolTransform invLutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, false);
-    bool customCmaxTU = false;
+    LUTsSubSymbolTransform lutsSubsymTrnsfm(supportVals, stateVars, numLuts, numPrvs, true);
     if(numLuts > 0) {
-        invLutsSubsymTrnsfm.decodeLUTs(reader);
-        if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
-            customCmaxTU = true;
-        }
+        lutsSubsymTrnsfm.encodeLUTs(writer, alphaProps, symbols, nullptr);
     }
 
     ContextSelector ctxSelector(stateVars);
@@ -378,64 +378,70 @@ size_t decodeTransformSubseqOrder2(const paramcabac::TransformedSubSeq &trnsfSub
                                 binParams);
 
     while (r.isValid()) {
-        // Decode subsymbols and merge them to construct symbols
-        uint64_t symbolValue = 0;
+        if (maxSize <= bitstream.size()) {
+            break;
+        }
 
+        // Split symbol into subsymbols and then encode subsymbols
+        const uint64_t origSymbol = r.get();
+        const int64_t signedSymbolValue = paramcabac::StateVars::getSignedValue(origSymbol, symbols->getWordSize());
+        const uint64_t symbolValue = abs(signedSymbolValue);
+        uint64_t subsymValToCode = 0;
+
+        uint32_t oss = outputSymbolSize;
         for (uint8_t s=0; s<stateVars.getNumSubsymbols(); s++) {
             const uint8_t lutIdx = (numLuts > 1) ? s : 0; // either private or shared LUT
             const uint8_t prvIdx = (numPrvs > 1) ? s : 0; // either private or shared PRV
 
+            subsymValToCode = subsymbols[s].subsymValue = (symbolValue>>(oss-=codingSubsymSize)) & subsymMask;
             subsymbols[s].subsymIdx = s;
+
             binParams[3] = ctxSelector.getContextIdxOrderGT0(s,
                                                              prvIdx,
                                                              subsymbols,
                                                              codingOrder);
 
-            if(customCmaxTU) {
-                subsymbols[s].lutNumMaxElems = invLutsSubsymTrnsfm.getNumMaxElemsOrder2(subsymbols,
-                                                                                        lutIdx,
-                                                                                        prvIdx);
-                binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
-            }
-            subsymbols[s].subsymValue = (reader.*func)(binParams);
-
             if(numLuts > 0) {
-                subsymbols[s].lutEntryIdx = subsymbols[s].subsymValue;
-                invLutsSubsymTrnsfm.invTransformOrder2(subsymbols, s, lutIdx, prvIdx);
+                subsymbols[s].lutEntryIdx = 0;
+                lutsSubsymTrnsfm.transformOrder2(subsymbols, s, lutIdx, prvIdx);
+                subsymValToCode = subsymbols[s].lutEntryIdx;
+                if(binID == paramcabac::BinarizationParameters::BinarizationId::TRUNCATED_UNARY) {
+                    binParams[0] = std::min((uint64_t) binarzationParams.getCMax(),subsymbols[s].lutNumMaxElems); // update cMax
+                }
             }
+
+            (writer.*func)(subsymValToCode, binParams);
 
             subsymbols[prvIdx].prvValues[1] = subsymbols[prvIdx].prvValues[0];
             subsymbols[prvIdx].prvValues[0] = subsymbols[s].subsymValue;
-
-            symbolValue = (symbolValue<<codingSubsymSize) | subsymbols[s].subsymValue;
         }
 
-        decodeSignFlag(reader, binID, symbolValue);
+        encodeSignflag(writer, binID, signedSymbolValue);
 
-        r.set(symbolValue);
         r.inc();
     }
 
-    payloadSizeUsed = reader.close();
+    writer.close();
 
-    decodedSymbols.swap(bitstream);
+    bitstream.flush(symbols);
 
-    return payloadSizeUsed;
+    return symbols->size(); // size of bitstream
 }
 
-size_t decodeTransformSubseq(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
-                             const unsigned int numEncodedSymbols,
-                             util::DataBlock* bitstream,
-                             util::DataBlock* const depSymbols) {
+size_t encodeTransformSubseq(const paramcabac::TransformedSubSeq &trnsfSubseqConf,
+                             util::DataBlock* symbols,
+                             util::DataBlock* const depSymbols,
+                             size_t maxSize) {
+
     switch(trnsfSubseqConf.getSupportValues().getCodingOrder()) {
         case 0:
-            return decodeTransformSubseqOrder0(trnsfSubseqConf, numEncodedSymbols, bitstream);
+            return encodeTransformSubseqOrder0(trnsfSubseqConf, symbols, maxSize);
         break;
         case 1:
-            return decodeTransformSubseqOrder1(trnsfSubseqConf, numEncodedSymbols, bitstream, depSymbols);
+            return encodeTransformSubseqOrder1(trnsfSubseqConf, symbols, depSymbols, maxSize);
         break;
         case 2:
-            return decodeTransformSubseqOrder2(trnsfSubseqConf, numEncodedSymbols, bitstream);
+            return encodeTransformSubseqOrder2(trnsfSubseqConf, symbols, maxSize);
         break;
         default:
             GABAC_DIE("Unknown coding order");
